More code_include test cleanup
Broke out code_include test content into files, this
allowed for much simplification on the tests
