+++ code_include/test_content/yourfile.rst
@@ -1,0 +1,13 @@
+How to Insult the English
+=========================
+
+:author: Pierre Devereaux
+
+A function to help insult those silly English knnnnnnniggets:
+
+.. code-include:: incfile.py
+    :lexer: python
+    :encoding: utf-8
+    :tab-width: 4
+    :start-line: 2
+    :end-line: 11
